Hayes: Allow using HTTPS URL as server address
The `Hayes.__init__` used to prepend "http://" to every server address
that does not start with "http://", which does not work for HTTPS urls.
Make it prepend the "http://" if there is no "://" in the given address.

In the odd case that address with "://" in middle is desired, the scheme
must be specified now.  This should not be a problem.

This fix is required to make Hayes work for shoop.io, since there the
ElasticSearch is accessed through HTTPS.

Refs SHOOPIO-215